services/auth/requirements.txt
blinker==1.6.3
click==8.1.7
Flask==3.0.0
Flask-MySQLdb==2.0.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
MarkupSafe==2.1.3
mysqlclient==2.2.0
PyJWT==2.8.0
python-dotenv==1.0.0
Werkzeug==3.0.0
